tests: reorganize some classes and modules
I moved the tetst events from SimulationTest to Helpers,
which I then moved to the global test module
