+++ src/test/java/fr/sncf/osrd/infra_state/RouteStateTest.java
@@ -1,6 +1,6 @@
 package fr.sncf.osrd.infra_state;
 
-import static fr.sncf.osrd.infra_state.Helpers.*;
+import static fr.sncf.osrd.Helpers.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
