+++ src/test/java/fr/sncf/osrd/infra_state/TVDSectionStateTest.java
@@ -1,6 +1,6 @@
 package fr.sncf.osrd.infra_state;
 
-import static fr.sncf.osrd.infra_state.Helpers.*;
+import static fr.sncf.osrd.Helpers.*;
 import static org.junit.jupiter.api.Assertions.fail;
 
 import fr.sncf.osrd.infra.InvalidInfraException;
